Bring Android UI up-to-scratch :-)
Use (modified) Android DoubleProgressBar [https://github.com/simao/Android-DoubleProgressBar]
for a double progress bar when en/decrypting directories. The original
project has licenced the code under the Apache License. It's been
modified a little to do exactly what we want (which was slightly
different to the original author), but the Android App now encrypts/
decrypts files/directories and displays the % complete as expected.