Update for backwards compatibility
Allow the user to specify which version of encrypt to encrypt their file
for. This can be set from the command line or the config file.

Will always default to the latest version, with all the
bells-and-whistles, but if version 2012.11 or earlier is requested, it
will actually fallback as far as possible.