One little enhancement
Add a command line option to not use the GUI (because I don't want it
popping up when I encrypt/decrypt on the fly with Vim).

The version number and file magic number have been updated to reflect
the new year, but as of yet undecided release month (hence the 00).