Work around for Windows 8
There appears to be a bug in libgcrypt which causes a segmentation
fault when attempting to use the AES/Rijndael cipher. The work around
currently removes this set of cipher from the list if it detects that
it's running on Windows 8 or newer. This bug does not appear on
Windows 7.